[CHORE] 폴더 구조 변경
