+++ src/main/java/kr/hhplus/be/server/api/coupon/dto/CouponResponse.java
@@ -1,0 +1,3 @@
+package kr.hhplus.be.server.api.coupon.dto;
+
+public record CouponResponse(Long couponId, String name, String discountType, int discountAmount) {}
