+++ src/main/java/kr/hhplus/be/server/api/point_history/dto/PointHistoryResponse.java
@@ -1,0 +1,3 @@
+package kr.hhplus.be.server.api.point_history.dto;
+
+public record PointHistoryResponse(Long userId, int amount, String type) {}
