--- src/main/java/kr/hhplus/be/server/domain/coupon/dto/CouponResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.domain.coupon.dto;
-
-public record CouponResponse(Long couponId, String name, String discountType, int discountAmount) {}
