--- src/main/java/kr/hhplus/be/server/domain/inventory/controller/InventoryController.java
@@ -1,14 +1,0 @@
-package kr.hhplus.be.server.domain.inventory.controller;
-
-import org.springframework.web.bind.annotation.*;
-import java.util.Map;
-
-@RestController
-@RequestMapping("/api/inventory")
-public class InventoryController {
-
-    @GetMapping("/{productId}")
-    public Map<String, Object> getInventory(@PathVariable Long productId) {
-        return Map.of("productId", productId, "stock", 50);
-    }
-}
