--- src/main/java/kr/hhplus/be/server/domain/inventory/dto/InventoryResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.domain.inventory.dto;
-
-public record InventoryResponse(Long productId, int stock) {}
