--- src/main/java/kr/hhplus/be/server/domain/payment/dto/PaymentResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.domain.payment.dto;
-
-public record PaymentResponse(Long orderId, String status, int paidAmount) {}
